Move custom components into separate files
