+++ packages/pipeline-editor/src/properties-panels/PropertiesPanel.tsx
@@ -16,20 +16,15 @@
 
 import { useState } from "react";
 
-import Form, {
-  ArrayFieldTemplateProps,
-  Field,
-  FieldTemplateProps,
-  UiSchema,
-  Widget,
-  utils,
-  WidgetProps,
-  FieldProps,
-  AjvError,
-} from "@rjsf/core";
+import Form, { UiSchema, Widget, AjvError } from "@rjsf/core";
 import styled from "styled-components";
 
-import { FileWidget } from "../CustomFormControls";
+import {
+  FileWidget,
+  CustomFieldTemplate,
+  ArrayTemplate,
+  CustomOneOf,
+} from "../CustomFormControls";
 
 export const Message = styled.div`
   margin-top: 14px;
@@ -52,250 +47,6 @@ interface Props {
   onFileRequested?: (options: any) => any;
   onPropertiesUpdateRequested?: (options: any) => any;
 }
-
-/**
- * React component that allows for custom add / remove buttons in the array
- * field component.
- */
-export const ArrayTemplate: React.FC<ArrayFieldTemplateProps> = (props) => {
-  return (
-    <div className={props.className}>
-      {props.items.map((item) => {
-        return (
-          <div key={item.key} className={item.className}>
-            {item.children}
-            <button
-              className="jp-mod-styled jp-mod-warn"
-              onClick={item.onDropIndexClick(item.index)}
-              disabled={!item.hasRemove}
-            >
-              {"Remove"}
-            </button>
-          </div>
-        );
-      })}
-      {props.uiSchema.pipeline_defaults?.map((item: any) => {
-        if (item === null || item === "") {
-          return undefined;
-        }
-        return (
-          <div className="array-pipelineDefaults form-control">
-            <div className="left">{item}</div>
-            <div className="right">(pipeline default)</div>
-          </div>
-        );
-      })}
-      {props.canAdd && (
-        <button
-          className="jp-mod-styled jp-mod-reject"
-          onClick={props.onAddClick}
-        >
-          {"Add"}
-        </button>
-      )}
-      {props.uiSchema.canRefresh && (
-        <button
-          className="jp-mod-styled jp-mod-reject"
-          style={{ marginLeft: "5px" }}
-          onClick={() =>
-            props.formContext?.onPropertiesUpdateRequested(
-              props.formContext.formData
-            )
-          }
-        >
-          {"Refresh"}
-        </button>
-      )}
-    </div>
-  );
-};
-
-/**
- * A custom oneOf field to handle the 2 custom oneOf cases that Elyra has:
- * 1. Each oneOf entry represents a different widget:
- *       Selecting the options based on widget type: assumes the value of the field
- *       is { value, widget } so that the widget type can be saved in the json.
- * 2. "inputpath": Each oneOf entry represents a different set object
- *       Selecting the options based on the default values of the given oneOf object.
- */
-export const CustomOneOf: Field = (props) => {
-  const { options, formData, registry } = props;
-  const findOption = (): any => {
-    // For inputpaths, expect a oneOf that has { value, option } for each entry
-    if ((props.schema as any).uihints?.["inputpath"]) {
-      for (const i in props.schema.oneOf ?? []) {
-        const properties: any = props.schema.oneOf?.[i];
-        if (
-          formData?.value === (properties as any).properties.value.default &&
-          formData?.option === (properties as any).properties.option.default
-        ) {
-          return i;
-        }
-      }
-      return 0;
-    }
-    // for other oneOfs, check for widget specified in the "value" uihints to match the saved widget
-    for (const i in options as any[]) {
-      const option = options[i];
-      if (option.properties.widget?.default === formData?.widget) {
-        return i;
-      }
-    }
-    return 0;
-  };
-  const [selectedOption, setSelectedOption] = useState(findOption());
-
-  const onOptionChange = (option: any) => {
-    const selectedOption = parseInt(option, 10);
-    const { rootSchema } = props.registry;
-    const newOption = utils.retrieveSchema(
-      options[selectedOption],
-      rootSchema,
-      formData
-    );
-
-    // If the new option is of type object and the current data is an object,
-    // discard properties added using the old option.
-    let newFormData = undefined;
-    if (
-      utils.guessType(formData) === "object" &&
-      (newOption.type === "object" || newOption.properties)
-    ) {
-      newFormData = Object.assign({}, formData);
-
-      const optionsToDiscard = options.slice();
-      optionsToDiscard.splice(selectedOption, 1);
-
-      // Discard any data added using other options
-      for (const option of optionsToDiscard) {
-        if (option.properties) {
-          for (const key in option.properties) {
-            if (newFormData.hasOwnProperty(key)) {
-              delete newFormData[key];
-            }
-          }
-        }
-      }
-    }
-    // Call getDefaultFormState to make sure defaults are populated on change.
-    let defaults;
-    try {
-      defaults = utils.getDefaultFormState(
-        options[selectedOption],
-        newFormData,
-        rootSchema
-      );
-    } catch {}
-    props.onChange(defaults);
-
-    setSelectedOption(parseInt(option, 10));
-  };
-
-  const SchemaField = registry.fields.SchemaField as React.FC<FieldProps>;
-  const { widgets } = registry;
-  const uiOptions = (utils.getUiOptions(props.uiSchema) ?? {}) as WidgetProps;
-  const Widget = utils.getWidget(
-    { type: "number" },
-    "select",
-    widgets
-  ) as React.FC<WidgetProps>;
-
-  const option = options[selectedOption] || null;
-  let optionSchema;
-
-  if (option) {
-    // If the subschema doesn't declare a type, infer the type from the
-    // parent schema
-    optionSchema = option.type
-      ? option
-      : Object.assign({}, option, { type: props.baseType });
-  }
-
-  const enumOptions = options.map((option: any, index: number) => ({
-    label: option.title || `Option ${index + 1}`,
-    value: index,
-  }));
-
-  return (
-    <div className="panel panel-default panel-body">
-      <div className="form-group">
-        <Widget
-          {...uiOptions}
-          id={`${props.idSchema.$id}${
-            props.schema.oneOf ? "__oneof_select" : "__anyof_select"
-          }`}
-          schema={{ type: "number", default: 0 }}
-          onChange={onOptionChange}
-          onBlur={props.onBlur}
-          onFocus={props.onFocus}
-          value={selectedOption}
-          options={{ enumOptions }}
-          registry={registry}
-        />
-      </div>
-
-      {option !== null && (
-        <SchemaField
-          {...props}
-          schema={optionSchema}
-          uiSchema={optionSchema.uihints}
-        />
-      )}
-    </div>
-  );
-};
-
-const CustomFieldTemplate: React.FC<FieldTemplateProps> = (props) => {
-  let children = props.children;
-  if (props.uiSchema["ui:field"] === "hidden") {
-    return <div />;
-  }
-  if (props.uiSchema["ui:readonly"]) {
-    children = (
-      <div style={{ paddingTop: "8px" }}>
-        {props.formData ?? props.schema.default}
-      </div>
-    );
-  }
-  const requiredError = props.required && props.formData === undefined;
-  const hasError = props.rawErrors || requiredError;
-  return (
-    <div
-      className={`${props.classNames} ${
-        props.schema.oneOf ? "field-oneOf" : ""
-      }`}
-      id={props.id}
-    >
-      {hasError && <div className="errorIndicator" />}
-      {props.schema.title !== undefined && props.schema.title !== " " ? (
-        <div
-          className={`label-header ${
-            props.uiSchema["ui:field"] === "header" ? "category-header" : ""
-          }`}
-        >
-          <label className="control-label" htmlFor={props.id}>
-            {`${props.schema.title}${props.required ? "*" : ""}`}
-          </label>
-          {props.schema.description && (
-            <div className="description-wrapper">
-              <div className="description-button">?</div>
-              <p
-                className={`field-description ${
-                  props.schema.title.length < 10 ? "short-title" : ""
-                }`}
-              >
-                {props.schema.description}
-              </p>
-            </div>
-          )}
-        </div>
-      ) : undefined}
-      {children}
-      {props.errors}
-      {requiredError && <li className="text-danger">is a required property</li>}
-    </div>
-  );
-};
 
 export function PropertiesPanel({
   data,
